+++ CHANGELOG.md
@@ -4,10 +4,18 @@
 Version 5.1-SNAPSHOT
 -------------
 
+ADDED: 
+
+- new dependency jobj-core in version 3.2.1
+
 CHANGED:
 
-- update of file-worker version to 5.0.2
-- update of test-objects dependency version to 5.0.1
+- update of parent version to 5
+- update of xmlbeans version to 3.1.0
+- update of file-worker version to 5.2.1
+- update of test-objects dependency version to 5.2
+- update of jobject-clone dependency version to 3.1.2
+- replaced obsolete dependency jobject-evaluate with new jobj-contract-verifier
 
 Version 5
 -------------